Added docker

+++ database/migrations/20240927153647_create_sessions_table.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE sessions (
-    session_token BIGINT PRIMARY KEY,
-    user_id BIGINT NOT NULL,
-    expires_at BIGINT NOT NULL,
+    session_token INTEGER PRIMARY KEY,
+    user_id INTEGER NOT NULL,
+    expires_at INTEGER NOT NULL,
     FOREIGN KEY (user_id) REFERENCES users(id) ON DELETE CASCADE
 );
